requirements.txt
Werkzeug==2.2.3
Flask==2.2.3
python-telegram-bot==13.15
